A game made with C# back in late 2017-2018.
This is an old code for a platform video game I made. The overall project turned out to be quite well although it does feel a bit scuffed. Unfortunately it's been so long that I do not remember the code base all too well to go in and optimize my code. Nonetheless, this game works fantastic.